Moved registrat and changed a check
